chore: remove unused modules
