--- RemootioAndroidAuto/automotive/build.gradle
@@ -1,65 +1,0 @@
-plugins {
-    id 'com.android.application'
-    id 'kotlin-android'
-    id 'org.jetbrains.kotlin.android'
-}
-
-def apikeyPropertiesFile = rootProject.file("apiKeys.properties")
-def apikeyProperties = new Properties()
-apikeyProperties.load(new FileInputStream(apikeyPropertiesFile))
-
-android {
-    namespace 'chamilton0.remootioandroidws.auto'
-    compileSdk 34
-
-    defaultConfig {
-        applicationId 'chamilton0.remootioandroidws'
-        minSdk 28
-        targetSdk 34
-        versionCode 5
-        versionName "0.2.0"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        applicationIdSuffix 'auto'
-    }
-
-    signingConfigs {
-        release {
-            storeFile file("../keystore.jks")
-            storePassword "Remootio"
-            keyAlias "key0"
-            keyPassword "Remootio"
-        }
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'),
-                    'proguard-rules.pro'
-            signingConfig signingConfigs.release
-        }
-    }
-    compileOptions {
-        targetCompatibility = JavaVersion.VERSION_17
-        sourceCompatibility = JavaVersion.VERSION_17
-    }
-    kotlinOptions {
-        jvmTarget = '17'
-    }
-}
-
-dependencies {
-    implementation "androidx.car.app:app:1.2.0"
-    implementation 'androidx.appcompat:appcompat:1.6.1'
-    implementation 'androidx.constraintlayout:constraintlayout-core:1.0.4'
-    implementation 'androidx.lifecycle:lifecycle-common-java8:2.7.0'
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation 'org.java-websocket:Java-WebSocket:1.5.5'
-    implementation 'androidx.core:core-ktx:1.12.0'
-    implementation 'org.apache.commons:commons-text:1.11.0'
-    implementation project(path: ':shared')
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
-}
