--- RemootioAndroidAuto/automotive/src/main/java/chamilton0/remootioandroidws/auto/RemootioService.kt
@@ -1,30 +1,0 @@
-package chamilton0.remootioandroidws.auto
-
-import android.content.Intent
-import android.content.pm.ApplicationInfo
-import androidx.car.app.CarAppService
-import androidx.car.app.R
-import androidx.car.app.Screen
-import androidx.car.app.Session
-import androidx.car.app.validation.HostValidator
-
-class RemootioService : CarAppService() {
-    override fun createHostValidator(): HostValidator {
-        return if (applicationInfo.flags and ApplicationInfo.FLAG_DEBUGGABLE != 0) {
-            HostValidator.ALLOW_ALL_HOSTS_VALIDATOR
-        } else {
-            HostValidator.Builder(applicationContext)
-                .addAllowedHosts(R.array.hosts_allowlist_sample)
-                .build()
-        }
-    }
-
-    //entry point of the application
-    override fun onCreateSession(): Session {
-        return object : Session() {
-            override fun onCreateScreen(intent: Intent): Screen {
-                return RemootioScreen(carContext)
-            }
-        }
-    }
-}
