--- RemootioAndroidAuto/automotive/src/main/res/drawable-v24/ic_fastfood_white_48dp.xml
@@ -1,24 +1,0 @@
-<!--
- Copyright (C) 2021 The Android Open Source Project
-
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--->
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="48dp"
-    android:height="48dp"
-    android:viewportWidth="48"
-    android:viewportHeight="48">
-  <path
-      android:pathData="M36.12,45.98h3.32c1.69,0 3.07,-1.27 3.26,-2.92L46,10.1L36,10.1L36,2h-3.94v8.1h-9.94l0.6,4.68c3.43,0.94 6.63,2.64 8.53,4.52 2.88,2.85 4.87,5.77 4.87,10.59L36.12,46v-0.02zM2,43.98L2,42h30.06v1.98c0,1.1 -0.9,2 -2.02,2L4.02,45.98c-1.12,0 -2.02,-0.91 -2.02,-2zM32.06,29.98c0,-16 -30.06,-16 -30.06,0h30.06zM2.04,34h30v4h-30z"
-      android:fillColor="#FFFFFF"/>
-</vector>
