--- RemootioAndroidAuto/automotive/src/main/res/values/themes.xml
@@ -1,4 +1,0 @@
-<resources>
-
-    <style name="Theme.Remootio" parent="android:Theme.Material.Light.DarkActionBar" />
-</resources>
