--- RemootioAndroidAuto/automotive/src/main/res/xml/automotive_app_desc.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<automotiveApp xmlns:android="http://schemas.android.com/apk/res/android">
-    <uses name="template" />
-</automotiveApp>
