--- RemootioAndroidAuto/automotive/src/test/java/chamilton0/remootioandroidws/ExampleUnitTest.java
@@ -1,17 +1,0 @@
-package chamilton0.remootioandroidws;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() {
-        assertEquals(4, 2 + 2);
-    }
-}
