+++ RemootioAndroidAuto/main/build.gradle
@@ -39,8 +39,6 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'com.google.android.material:material:1.11.0'
     implementation 'androidx.preference:preference-ktx:1.2.1'
-    implementation project(path: ':shared')
-    implementation project(path: ':automotive', configuration: 'default')  // Required for the build
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
@@ -52,7 +50,6 @@ dependencies {
     implementation 'org.java-websocket:Java-WebSocket:1.5.5'
     implementation 'androidx.core:core-ktx:1.12.0'
     implementation 'org.apache.commons:commons-text:1.11.0'
-    implementation project(path: ':shared')
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
