+++ RemootioAndroidAuto/settings.gradle
@@ -6,6 +6,4 @@ dependencyResolutionManagement {
     }
 }
 rootProject.name = "RemootioAndroidWS"
-include ':automotive'
 include ':main'
-include ':shared'
